+++ flask_quiz/static/custom_style.css
@@ -1,0 +1,4 @@
+.jumbotron {
+    background-color: #f4511e;
+    color: #fff;
+}
