+++ flask_quiz/templates/highscore_page.html
@@ -1,0 +1,26 @@
+{% extends "bootstrap/base.html" %}
+{% block styles %}
+{{super()}}
+<link rel="stylesheet"
+      href="{{url_for('.static', filename='custom_style.css')}}">
+{% endblock %}
+{% block title %}Quiz{% endblock %}
+{% block content %}
+<div class="jumbotron text-center">
+    <p>Highscore</p>
+    <div>
+        <table class="table text-left">
+            <tr>
+                <th scope="col" width="25%">completion date time</th>
+                <th scope="col" width="75%">name</th>
+            </tr>
+            {% for user in users %}
+            <tr>
+                <th scope="col">{{user.completion_date_time}}</th>
+                <td>{{user.name}}</td>
+            </tr>
+            {% endfor %}
+        </table>
+    </div>
+</div>
+{% endblock %}
