+++ requirements.txt
@@ -10,3 +10,4 @@ SQLAlchemy==1.3.18
 waitress==1.4.4
 Werkzeug==1.0.1
 wordcloud~=1.7.0
+WTForms~=2.3.1
