--- static/custom_style.css
@@ -1,4 +1,0 @@
-.jumbotron {
-    background-color: #f4511e;
-    color: #fff;
-}
