--- templates/naming_page.html
@@ -1,18 +1,0 @@
-{% extends "bootstrap/base.html" %}
-{% block styles %}
-{{super()}}
-<link rel="stylesheet"
-      href="{{url_for('.static', filename='custom_style.css')}}">
-{% endblock %}
-{% block title %}Quiz{% endblock %}
-{% block content %}
-<div class="jumbotron text-center">
-    <p>Enter your name:</p>
-    <form class="form-inline" action="/user/{{id}}/name" method="POST">
-        <div class="input-group-append">
-                {{ form.name(class="form-control") }}
-                {{ form.submit(class="btn btn-primary") }}
-        </div>
-    </form>
-</div>
-{% endblock %}
