--- templates/question_page.html
@@ -1,28 +1,0 @@
-{% extends "bootstrap/base.html" %}
-{% block styles %}
-{{super()}}
-<link rel="stylesheet"
-      href="{{url_for('.static', filename='custom_style.css')}}">
-{% endblock %}
-{% block title %}Quiz{% endblock %}
-{% block content %}
-
-<div class="jumbotron text-center">
-    {% if is_last_answer_correct == True %}
-        <div class="alert alert-success">
-            <strong>Right! </strong>Your last answer was correct.
-        </div>
-    {% elif is_last_answer_correct == False %}
-        <div class="alert alert-danger">
-            <strong>Wrong! </strong>Your last answer wasn't correct.
-        </div>
-    {% endif %}
-    <p>{{question.id}}. {{question.text}}</p>
-    <form action="/user/{{id}}/answer" method="POST">
-        <div class="form-group">
-                {{ form.answer(class="form-control") }}
-                {{ form.submit(class="btn btn-primary") }}
-        </div>
-    </form>
-</div>
-{% endblock %}
